feat: split api between vanilla and react
