+++ src/react.ts
@@ -1,3 +1,6 @@
 export { useAtomsLogger } from './react/use-atoms-logger.js';
-export { bindAtomsLoggerToStore } from './vanilla/bind-atoms-logger-to-store.js';
+export {
+  bindAtomsLoggerToStore,
+  isAtomsLoggerBoundToStore,
+} from './vanilla/bind-atoms-logger-to-store.js';
 export type { AtomsLoggerOptions } from './vanilla/types/atoms-logger.js';
