+++ src/vanilla.ts
@@ -1,0 +1,5 @@
+export {
+  bindAtomsLoggerToStore,
+  isAtomsLoggerBoundToStore,
+} from './vanilla/bind-atoms-logger-to-store.js';
+export type { AtomsLoggerOptions } from './vanilla/types/atoms-logger.js';
