+++ src/vanilla/bind-atoms-logger-to-store.ts
