+++ src/vanilla/callbacks/on-atom-garbage-collected.ts
