+++ src/vanilla/callbacks/on-atom-mounted.ts
