+++ src/vanilla/callbacks/on-atom-state-map-set.ts
