+++ src/vanilla/log-atom-event/add-atom-to-logs.ts
