+++ src/vanilla/log-atom-event/add-dash-to-logs.ts
