+++ src/vanilla/log-atom-event/add-to-logs.ts
