+++ src/vanilla/log-atom-event/event-log-pipeline.ts
