+++ src/vanilla/log-atom-event/log-event.ts
