+++ src/vanilla/log-atom-event/log-pipeline.ts
