+++ src/vanilla/log-atom-event/transaction-log-pipeline.ts
