+++ src/vanilla/transactions/end-transaction.ts
