+++ src/vanilla/transactions/stop-end-transaction-debounce.ts
