+++ src/vanilla/utils/has-atom-custom-write-method.ts
