+++ src/vanilla/utils/logger-options-to-state.ts
