+++ src/vanilla/utils/parse-owner-stack.ts
