+++ src/vanilla/utils/stringify-value.ts
